Python/requirements.txt
requests
re
pyserial
SpeechRecognition
pyaudio
python-dotenv